### 2025-09-03 #8926 @editorialbot
[PRE REVIEW]: modepy: Basis Functions, Interpolation, and Quadrature (not just) for Finite Elements
**Submitting author:** @alexfikl (<a href="http://orcid.org/0000-0002-0552-5936">Alexandru Fikl</a>)
**Repository:** https://github.com/inducer/modepy
**Branch with paper.md** (empty if default branch): joss
**Version:** v2025.1
