### 2025-09-03 #8926 @danielskatz
@alexfikl  - thanks for your submission.

First, you should remove the first two lines in your license file so that it is a standard MIT license, without additional text.

Second, your paper is too long. JOSS papers are recommended to be 250-1000 words, while yours is 2651.  Your paper is also missing the References header at the end that would appear before the references in the generated PDF.

Please feel free to use the command `@editorialbot generate pdf` to make a new PDF and the command `@editorialbot check repository` to check a bunch of stuff in the repository, including the word count.

In addition, you could work on the possibly missing DOI that editorialbot suggests, but note that it may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again.

editorialbot commands need to be the first entry in a new comment.

Alternatively, if you want to keep the papers content, I can mark the paper as withdrawn, and you can submit it to another venue where a longer paper would be appropriate.

Finally, I note that I have an institutional conflict of interest with @inducer and @a-alveyblanc, so once we get to the point of having a submission that JOSS is ready to process, I will ask another track editor to take over on assignments, decisions, etc.
