### 2025-09-04 ea00e17
Fix recursivenodes reference and add doi
