+++ doc/paper/paper.bib
@@ -115,8 +115,12 @@
   },
   author        = {Tobin Isaac},
   year          = 2020,
-  eprint        = {2002.09421v1},
-  eprinttype    = {arXiv},
+  journal       = {{SIAM} Journal on Scientific Computing},
+  publisher     = {Society for Industrial \& Applied Mathematics {(SIAM)}},
+  volume        = 42,
+  pages         = {A4046--A4062},
+  doi           = {10.1137/20m1321802},
+  issue         = 6,
 }
 
 @article{Jaskowiec2021,
