### 2025-09-04 c1d5027
Add References section for bibliography
