+++ doc/paper/paper.md
@@ -382,3 +382,5 @@ Multidisciplinary University Research Initiatives (MURI) Program, under Grant
 Number *N00014-16-1-2617*. A. Klöckner was supported by the US National Science
 Foundation under award number DMS-2410943, and by the US Department of Energy
 under award number DE-NA0003963.
+
+# References
