### 2025-09-04 #8926 @danielskatz
> > Second, your paper is too long. JOSS papers are recommended to be 250-1000 words, while yours is 2651.
>
> We will attempt to shorten the paper to a more acceptable length, but it may take a week or two. Is it ok for the pre-review to remain open until then or is a re-submission more appropriate?

yes, that's fine

> Also, is the 1000 word limit a hard limit? Would a paper with 1000-1300 words be acceptable as well?

slightly over may be ok, but it depends a bit on if editors and reviewers see things that could be in the repo or documentation instead, given that a JOSS paper is not a theory paper but a software paper.
