### 2025-09-05 adeb751
docs: update index.rst links and match readme
