+++ doc/index.rst
@@ -7,14 +7,11 @@ Welcome to modepy's documentation!
 simplices (i.e. triangles and tetrahedra) and tensor products of simplices
 (i.e. squares, cubes, prisms, etc.). These are a key building block for
 high-order unstructured discretizations, as often used in a finite element
-context. It closely follows the approach taken in the book
+context.
 
-  Hesthaven, Jan S., and Tim Warburton. "Nodal Discontinuous Galerkin Methods:
-  Algorithms, Analysis, and Applications". 1st ed. Springer, 2007.
-  `Book web page <http://nudg.org>`_
-
-The basic objects that :mod:`modepy` manipulates are functions on a simplex.
-For example, it supplies an orthonormal basis on triangles (shown here) and tetrahedra.
+The basic objects that :mod:`modepy` manipulates are functions on a shape (or
+reference domain). For example, it supplies an orthonormal basis on triangles
+(shown below).
 
 .. image:: images/pkdo-2d.png
     :width: 100%
@@ -23,6 +20,18 @@ For example, it supplies an orthonormal basis on triangles (shown here) and tetr
 
 The file that created this plot is included in the :mod:`modepy` distribution
 as :download:`examples/plot-basis.py <../examples/plot-basis.py>`.
+
+Its roots closely followed the approach taken in the Hesthaven and Warburton
+book [HesthavenWarburton2007]_, but much has been added beyond that basic
+functionality.
+
+.. [HesthavenWarburton2007] J. S. Hesthaven and T. Warburton (2007).
+    *Nodal Discontinuous Galerkin Methods: Algorithms, Analysis, and Applications*
+    (1st ed.).
+    `doi:10.1007/978-0-387-72067-8 <https://doi.org/10.1007/978-0-387-72067-8>`__.
+
+Example
+-------
 
 Here's an idea of code that uses :mod:`modepy`:
 
@@ -34,10 +43,10 @@ This file is included in the :mod:`modepy` distribution as
 modepy around the web
 ---------------------
 
-* `Wiki home page <http://wiki.tiker.net/ModePy>`_
-* `Source code <http://github.com/inducer/modepy>`_
-* `Bug tracker <http://github.com/inducer/modepy/issues>`_
-* `Package index <http://pypi.python.org/pypi/modepy>`_
+* `Documentation <https://documen.tician.de/modepy>`__.
+* `Source code <https://github.com/inducer/modepy>`__.
+* `Bug tracker <https://github.com/inducer/modepy/issues>`__.
+* `PyPI <https://pypi.org/project/modepy/>`__.
 
 Contents
 ========
