### 2025-09-05 5f1bfa4
docs: use https in README and match index.rst
