### 2025-09-05 6e0b843
docs: use prism example in index front-page
