### 2025-09-07 08e7735
docs: add link to nudg github repo
