### 2025-09-07 cae5a7d
docs: add link to nudg github repo
