+++ README.rst
@@ -54,6 +54,7 @@ Its roots closely followed the approach taken in the book
     *Nodal Discontinuous Galerkin Methods: Algorithms, Analysis, and Applications*
     (1st ed.).
     `doi:10.1007/978-0-387-72067-8 <https://doi.org/10.1007/978-0-387-72067-8>`__.
+    `(source code) <https://github.com/tcew/nodal-dg>`__
 
 but much has been added beyond that basic functionality.
 
