+++ .github/workflows/ci.yml
@@ -17,7 +17,7 @@ jobs:
         runs-on: ubuntu-latest
         steps:
         -   uses: actions/checkout@v5
-        -   uses: actions/setup-python@v5
+        -   uses: actions/setup-python@v6
         -   name: "Main Script"
             run: |
                 pip install ruff
@@ -27,7 +27,7 @@ jobs:
         runs-on: ubuntu-latest
         steps:
         -   uses: actions/checkout@v5
-        -   uses: actions/setup-python@v5
+        -   uses: actions/setup-python@v6
             with:
                 python-version: '3.x'
         -   name: "Main Script"
@@ -54,7 +54,7 @@ jobs:
         steps:
         -   uses: actions/checkout@v5
         -
-            uses: actions/setup-python@v5
+            uses: actions/setup-python@v6
             with:
                 python-version: ${{ matrix.python-version }}
         -   name: "Main Script"
@@ -69,7 +69,7 @@ jobs:
         steps:
         -   uses: actions/checkout@v5
         -
-            uses: actions/setup-python@v5
+            uses: actions/setup-python@v6
             with:
                 python-version: '3.x'
         -   name: "Main Script"
@@ -87,7 +87,7 @@ jobs:
         steps:
         -   uses: actions/checkout@v5
         -
-            uses: actions/setup-python@v5
+            uses: actions/setup-python@v6
             with:
                 python-version: '3.x'
         -   name: "Main Script"
