### 2025-09-12 94fbac1
Remove example and trim down overview
