+++ doc/paper/paper.md
@@ -146,16 +146,11 @@ to any array type (including GPU or other high-performance arrays) as long as
 
 # Overview
 
-The high-level concepts available in `modepy` are shapes (i.e the geometry of a
+The high-level concepts available in `modepy` are shapes (i.e. a reference
 domain), modes (i.e. the basis functions), and nodes (i.e. the degrees of
 freedom). These are implemented in a user-extensible fashion using the
 `singledispatch` mechanism, with inspiration taken from common idiomatic usage
-in Julia [@Bezanson2017]. Given a function space `space` on a `shape`, a set of
-nodal degrees of freedom can be obtained via
-`edge_clustered_nodes_for_space(space, shape)`, with an implementation selected
-based on the type of `space`. While the example shows that multiple dispatch
-could be beneficial, we have decided to forgo it to avoid potential soundness
-issues [@Julia2018].
+in Julia [@Bezanson2017]. 
 
 ## Shapes
 
@@ -163,126 +158,68 @@ The geometry of a reference element is described in `modepy` by the `Shape`
 class. Built-in support exists for `Simplex` and `Hypercube` geometries,
 encompassing the commonly used interval, triangle, tetrahedron, quadrilateral,
 and hexahedral shapes (see \autoref{FigureSimplices}). `TensorProductShape` can
-be used to compose additional shapes, such as prisms, that are useful in
-specific applications and sometimes generated by meshing software (such as
-`gmsh` [@Geuzaine2009]). An example that constructs a prism and performs some
-standard operations on it is given in the next section.
+be used to compose additional shapes, that are useful in specific applications
+and sometimes generated by meshing software (such as `gmsh` [@Geuzaine2009]). 
 
 ![Domains corresponding to the one-, two-, and three-dimensional simplices.](images/simplices.png){#FigureSimplices width="80%"}
 
-To support a wide array of applications, shapes can be queried and manipulated
-in several ways. The `faces_for_shape` function can be used to return a list of
-faces, which are shapes themselves, that allow access to domain normals and
-mappings back to the reference element.
-
 ## Modes and Spaces
 
-To perform calculus operations, each reference element can be equipped with
-function spaces described by the `FunctionSpace` class.
-These represent a finite-dimensional space of functions $\phi_i: D \to
-\mathbb{R}$, where $D$ is the reference element domain, but make no specific
-choice of basis. The function spaces mirror the shape definitions and are
-intended to work in tandem. In fact, the `space_for_shape` function provides a
-default choice of space for each shape. Predefined choices include the `PN`
-space, containing polynomials of total degree at most `order`, and the `QN`
-space, containing polynomials of maximum degree at most `order`. As with
-shapes, these spaces can also be combined using `TensorProductSpace` to
-describe functional spaces for other domains or in higher dimensions.
+To perform calculus operations, each reference element can be equipped with a
+function space described by the `FunctionSpace` class. These represent a
+finite-dimensional space of functions $\phi_i: D \to \mathbb{R}$, where $D$ is
+the reference element domain, and no specific choice of basis.  Predefined
+choices include the `PN` space, containing polynomials of total degree at most
+`order`, and the `QN` space, containing polynomials of maximum degree at most
+`order`. As with shapes, these spaces can be combined using
+`TensorProductSpace`.
 
-For calculations, each space can be equipped with a default basis using the
-`basis_for_space` function. `modepy` provides two special types of basis
-functions: orthonormal (accessed through `orthonormal_basis_for_space`) and
[…]
