+++ doc/paper/paper.md
@@ -41,108 +41,67 @@ bibliography: paper.bib
 
 # Summary
 
-`modepy` provides a means for constructing geometric shapes for reference
-elements, equipping them with appropriate approximation spaces, and numerically
-performing calculus operations (derivatives, integrals) on those spaces.
+`modepy` is a Python library for defining reference elements, equipping them
+with appropriate approximation spaces, and numerically performing calculus
+operations (derivatives, integrals) on those spaces. It is written in pure
+modern Python 3, offering comprehensive type annotations and full documentation
+with minimal runtime dependencies (NumPy being a primary dependency). 
 
 `modepy` focuses on high-order accuracy -- given an element size $h$, this
 refers to the asymptotic decay of the approximation error as $O(h^n)$, for $n
-\ge 3$, assuming sufficient smoothness of the solution being approximated. If
-we consider a problem in $d$ dimensions, the number of unknowns scales as
-$O(h^{-d})$. In the absence of high-order approximation, say at $n=1$, even
-modest accuracy increases (e.g. an additional significant digit) can incur
-significant cost (a $1000\times$ increase in the problem size in 3D).
-Therefore, if accuracy is desired at manageable cost, high-order methods are
-crucial.
-
-A popular approach for accurate approximation of functions on geometrically
-complex domains is the use of *unstructured discretizations*, which represent
-the geometry as a typically disjoint union (a "mesh") of primitive geometric
-shapes, most often simplices and quadrilaterals. Given the means to perform
-calculus operations on these *reference elements* and mapping functions from
-them to the *global* elements, calculus operations become available on the
-entire domain. Notably, one can approximate the reference-to-global coordinate
-mapping function itself with the same high-order machinery. Altogether, this
-paves the way for the solution of integral and (partial) differential
+\ge 3$, assuming sufficient smoothness of the solution being approximated -- at
+manageable computational cost. These capabilities enable High-Performance
+Computing (HPC) approaches to solving integral and (partial) differential
 equations. Those, in turn, can be used to model many phenomena in the physical
-world, including fluid flow, electromagnetism, and solid mechanics. The finite
-element method (FEM) is a popular example of this, including its continuous and
-discontinuous flavors. Further examples include collocation, spectral, and
-Nyström methods.
+world, including fluid flow, electromagnetism, and solid mechanics. A popular
+example of this is the finite element method (FEM), including its continuous
+and discontinuous flavors, but general collocation, spectral, and Nyström
+methods are also supported. As such, `modepy` has been used to construct FEM
+solvers [@Grudge; @PyNucleus] and integral equation solvers [@Pytential] that
+run on both CPUs and GPUs.
 
-`modepy` has been used to construct FEM solvers [@Grudge; @PyNucleus] and
-integral equation solvers [@Pytential]. It is written in pure modern Python 3,
-offering comprehensive type annotations and full documentation with minimal
-runtime dependencies, of which numpy is perhaps the most significant. Versions
-going back to as early as 2013 offer broad compatibility with older versions of
-Python, including Python 2.7.
-
-`modepy` is licensed under the MIT license and available on Github at
-<https://github.com/inducer/modepy/>.
+`modepy` is licensed under the MIT license and available on GitHub at
+<https://github.com/inducer/modepy/>. Versions dating back to 2013 maintain
+broad compatibility with older versions of Python, including Python 2.7,
+ensuring broad accessibility.
 
 # Statement of need
 
 High-order accurate calculus operations on unstructured discretizations are
[…]
