### 2025-09-19 0913f0e
Another shortening pass
