+++ doc/paper/paper.md
@@ -43,39 +43,37 @@ bibliography: paper.bib
 
 `modepy` is a Python library for defining reference elements, equipping them
 with appropriate approximation spaces, and numerically performing calculus
-operations (derivatives, integrals) on those spaces. It is written in pure
-modern Python 3, offering comprehensive type annotations and full documentation
-with minimal runtime dependencies (NumPy being a primary dependency). 
+operations (derivatives, integrals) on those spaces. It is written in pure,
+type-annotated Python 3, offering comprehensive documentation
+and minimal runtime dependencies (mainly NumPy).
 
-`modepy` focuses on high-order accuracy -- given an element size $h$, this
-refers to the asymptotic decay of the approximation error as $O(h^n)$, for $n
-\ge 3$, assuming sufficient smoothness of the solution being approximated -- at
-manageable computational cost. These capabilities enable High-Performance
-Computing (HPC) approaches to solving integral and (partial) differential
-equations. Those, in turn, can be used to model many phenomena in the physical
-world, including fluid flow, electromagnetism, and solid mechanics. A popular
-example of this is the finite element method (FEM), including its continuous
-and discontinuous flavors, but general collocation, spectral, and Nyström
-methods are also supported. As such, `modepy` has been used to construct FEM
-solvers [@Grudge; @PyNucleus] and integral equation solvers [@Pytential] that
-run on both CPUs and GPUs.
+`modepy` focuses on high-order accuracy --- given an element size $h$, this refers to
+the asymptotic decay of the approximation error as $O(h^n)$, for $n \ge 3$, assuming
+sufficient smoothness of the solution being approximated. For a problem in $d$
+dimensions, the number of unknowns scales as $O(h^{-d})$. Therefore, if accuracy is
+desired at manageable cost, high-order methods are crucial.
 
-`modepy` is licensed under the MIT license and available on GitHub at
-<https://github.com/inducer/modepy/>. Versions dating back to 2013 maintain
-broad compatibility with older versions of Python, including Python 2.7,
-ensuring broad accessibility.
+A popular approach (e.g. in the finite element method, or FEM) for accurate
+approximation of functions on geometrically complex domains is the use of *unstructured
+discretizations*, which represent the geometry as a typically disjoint union (a "mesh")
+of primitive geometric shapes, most often simplices and quadrilaterals. Given the means
+to perform calculus operations on these *reference elements* and mapping functions
+from them to the *global* elements, calculus operations become available on the entire
+domain.  These primitives are chiefly useful in the numerical solution of integral and
+(partial) differential equations. Additional applications include computer graphics,
+CAD, and robotics. Those, in turn, can be used to model many physical phenomena,
+including fluid flow, electromagnetism, and solid mechanics. `modepy` has been used to
+construct FEM solvers [@Grudge; @PyNucleus] and integral equation solvers [@Pytential]
+that run on both CPUs and GPUs.
 
 # Statement of need
 
-High-order accurate calculus operations on unstructured discretizations are
-crucial to differential and integral equations solvers, as well as computer
-graphics and Computer-Aided Design (CAD) applications. This functionality is
-often embedded in an ad-hoc manner in larger codes, restricting scope and
-reusability. `modepy` addresses this need by providing a reusable,
-generalizable, and composable implementation.
+`modepy`'s functionality, as outlined above, is often embedded in an ad-hoc manner
+in larger codes, restricting scope and reusability. `modepy` addresses this need by
[…]
