### 2025-09-19 936934a
feat(paper): small rephrasing and reformatting
