### 2025-09-19 576cc3e
feat(paper): clean up and fix up bibliography
