+++ doc/paper/paper.bib
@@ -1,345 +1,347 @@
 @article{Bezanson2017,
-    title = {Julia: {A} Fresh Approach to Numerical Computing},
-    author = {Jeff Bezanson and Alan Edelman and Stefan Karpinski and Viral B.
-              Shah},
-    year = 2017,
-    journal = {{SIAM} Review},
-    publisher = {Society for Industrial \& Applied Mathematics {(SIAM)}},
-    volume = 59,
-    pages = {65--98},
-    doi = {10.1137/141000671},
-    issue = 1,
+  title         = {Julia: {A} Fresh Approach to Numerical Computing},
+  author        = {Jeff Bezanson and Alan Edelman and Stefan Karpinski and Viral B. Shah},
+  year          = 2017,
+  journal       = {{SIAM} Review},
+  publisher     = {Society for Industrial \& Applied Mathematics {(SIAM)}},
+  volume        = 59,
+  pages         = {65--98},
+  doi           = {10.1137/141000671},
+  issue         = 1,
 }
 
 @book{Brenner2007,
-    title = {The Mathematical Theory of Finite Element Methods},
-    author = {Susanne Brenner and Ridgway Scott},
-    year = 2007,
-    publisher = {Springer Science \& Business Media},
-    isbn = 9780387759333,
+  title         = {The Mathematical Theory of Finite Element Methods},
+  author        = {Susanne Brenner and Ridgway Scott},
+  year          = 2007,
+  publisher     = {Springer Science \& Business Media},
+  isbn          = 9780387759333,
 }
 
 @article{Dubiner1991,
-    title = {Spectral Methods on Triangles and Other Domains},
-    author = {M. Dubiner},
-    year = 1991,
-    journal = {Journal of Scientific Computing},
-    publisher = {Springer Science and Business Media {LLC}},
-    volume = 6,
-    pages = {345--390},
-    doi = {10.1007/bf01060030},
-    issue = 4,
+  title         = {Spectral Methods on Triangles and Other Domains},
+  author        = {M. Dubiner},
+  year          = 1991,
+  journal       = {Journal of Scientific Computing},
+  publisher     = {Springer Science and Business Media {LLC}},
+  volume        = 6,
+  pages         = {345--390},
+  doi           = {10.1007/bf01060030},
+  issue         = 4,
 }
 
 @software{FIAT,
-    title = {{Fiat}: 2025.4.0},
-    author = { Mikl\'{o}s Homolya and Pablo Brubeck and David A. Ham and Robert
-              Kirby and Lawrence Mitchell and Jan Blechta and Marie E Rognes and
-              Garth N. Wells and Anders Logg and cyruscycheng21 and Thomas H.
-              Gibson and Johannes Ring and Ivan Yashchuk and Justincrum and
-              k-b-oelgaard and ksagiyam and Chris Richardson and Nico Schl\"{o}
-              mer and Reuben W. Nixon-Hill and Lizao Li and Aslak Bergersen and
-              Connor Ward and Nicholas Barton and Colin J Cotter and Patrick E.
-              Farrell and Jack S. Hale and FAznaran and Matthew Knepley and
-              Florian Rathgeber and Matthew Scroggs },
-    year = 2025,
-    month = apr,
-    publisher = {Zenodo},
-    doi = {10.5281/zenodo.15302339},
-    url = {https://doi.org/10.5281/zenodo.15302339},
-    version = {2025.4.0},
+  title         = {{FIAT}: 2025.4.0},
+  author        = {
+    Mikl\'{o}s Homolya and Pablo Brubeck and David A. Ham and Robert Kirby and
+    Lawrence Mitchell and Jan Blechta and Marie E Rognes and Garth N. Wells and
[…]
