### 2025-09-19 #8926 @alexfikl
@danielskatz We have fixed all the initially mentioned issues (around the DOIs and the license text) and shortened the manuscript to a more manageable size. Let me know if there is anything else needed to move forward.
