### 2025-09-28 345c7cb
Fix ruff: extraneous str()
