### 2025-10-05 f1f72d8
fix: some newly surfaced typos
