### 2025-10-19 7daa176
feat(typing): small type updates
