### 2025-10-19 58c2fec
feat(typing): small type updates
