### 2025-10-20 #8926 @alexfikl
> [@alexfikl](https://github.com/alexfikl) My apologies for the delay, let’s get this started.

Absolutely no problem at all!

> Are there any reviewers you’d like to recommend from here: [reviewers.joss.theoj.org/lookup](https://reviewers.joss.theoj.org/lookup)?

We took a quick look and would recommend someone along the lines of [Andrew Giuliani](https://reviewers.joss.theoj.org/reviewers/20370), [Santiago Ospina De Los Ríos](https://reviewers.joss.theoj.org/reviewers/4087), [Kris Rowe](https://reviewers.joss.theoj.org/reviewers/2072), or [Mikael Mortensen](https://reviewers.joss.theoj.org/reviewers/433). We've looked for reviewers who work with FEM type methods, since they would be the main consumers of a library like `modepy`.

> In the meantime, I had a chance to look over the submission. Your `JOSS` branch is 4 commits behind `main`; could you please rebase?

Done!

> I can't not find a `CONTRIBUTING` file or guidelines. Could you make sure that the `CONTRIBUTING` file includes a section on how to run the tests?

Thank you for the recommendations, they're very helpful. I'll add a simple `CONTRIBUTING` file in there to explain the main workflow / requirements.
