### 2025-10-20 f48e696
feat: add a CONTRIBUTING.md file
