### 2025-10-20 36f023e
fix: improve wording in CONTRIBUTING.md
Co-authored-by: Andreas Klöckner <inform@tiker.net>
