### 2025-10-20 524b4b0
feat: add conda-forge to CONTRIBUTING.md
