### 2025-10-21 d880c12
Add a CONTRIBUTING.md file (#145)
* feat: add a CONTRIBUTING.md file

* fix: improve wording in CONTRIBUTING.md

Co-authored-by: Andreas Klöckner <inform@tiker.net>

* feat: add conda-forge to CONTRIBUTING.md

---------

Co-authored-by: Andreas Klöckner <inform@tiker.net>
