### 2025-10-23 #8926 @capitalaslash
Yes, I am available to review this package.
