### 2025-10-24 inducer/modepy#146 @capitalaslash
mention testing functionalities in documentation
The documentation does not refer to `pytest` anywhere.

I managed to run the testsuite easily myself, but I would prefer to see some instructions on how to test the library in the documentation, maybe at the end of the Installation part?

ref. https://github.com/openjournals/joss-reviews/issues/9294
