### 2025-10-27 f97d6d3
docs: refer to CONTRIBUTING.md for extra install info
