### 2025-10-27 inducer/modepy#146 @alexfikl
We do have a (recent) `CONTRIBUTING.md` file that should explain all the requirements and testing setup:
https://github.com/inducer/modepy/blob/main/CONTRIBUTING.md

I added a mention to it in the docs in #147. Would that be sufficient?
