### 2025-11-09 #9294 @felipefr
Sorry for the late reply. I'm going to review it tomorrow morning (CET).

Felipe Figueredo

On Sun, Nov 9, 2025, 12:20 PM Nikoleta Glynatsi ***@***.***>
wrote:

> *Nikoleta-v3* left a comment (openjournals/joss-reviews#9294)
> <https://github.com/openjournals/joss-reviews/issues/9294#issuecomment-3508023418>
>
> Thank you, @capitalaslash <https://github.com/capitalaslash> 🙏🏻 !!!
> @felipefr <https://github.com/felipefr> did you get a chance to go over
> the submission?
>
