+++ .github/workflows/ci.yml
@@ -16,7 +16,7 @@ jobs:
         name: Ruff
         runs-on: ubuntu-latest
         steps:
-        -   uses: actions/checkout@v5
+        -   uses: actions/checkout@v6
         -   uses: actions/setup-python@v6
         -   name: "Main Script"
             run: |
@@ -26,7 +26,7 @@ jobs:
     basedpyright:
         runs-on: ubuntu-latest
         steps:
-        -   uses: actions/checkout@v5
+        -   uses: actions/checkout@v6
         -   uses: actions/setup-python@v6
             with:
                 python-version: '3.x'
@@ -42,7 +42,7 @@ jobs:
         name: Typos
         runs-on: ubuntu-latest
         steps:
-        -   uses: actions/checkout@v5
+        -   uses: actions/checkout@v6
         -   uses: crate-ci/typos@master
 
     pytest:
@@ -52,7 +52,7 @@ jobs:
             matrix:
                 python-version: ['3.10', '3.x']
         steps:
-        -   uses: actions/checkout@v5
+        -   uses: actions/checkout@v6
         -
             uses: actions/setup-python@v6
             with:
@@ -67,7 +67,7 @@ jobs:
         name: Documentation
         runs-on: ubuntu-latest
         steps:
-        -   uses: actions/checkout@v5
+        -   uses: actions/checkout@v6
         -
             uses: actions/setup-python@v6
             with:
@@ -85,7 +85,7 @@ jobs:
         name: Examples Py3
         runs-on: ubuntu-latest
         steps:
-        -   uses: actions/checkout@v5
+        -   uses: actions/checkout@v6
         -
             uses: actions/setup-python@v6
             with:
@@ -106,7 +106,7 @@ jobs:
         name: Tests for downstream project ${{ matrix.downstream_project }}
         runs-on: ubuntu-latest
         steps:
-        -   uses: actions/checkout@v5
+        -   uses: actions/checkout@v6
         -   name: "Main Script"
             env:
                 DOWNSTREAM_PROJECT: ${{ matrix.downstream_project }}
