### 2026-01-11 d6a7309
Disable RUF067 for now
