+++ pyproject.toml
@@ -91,6 +91,7 @@ extend-ignore = [
     "E242", # tab after comma
     "E402", # module level import not at the top of file
     "N818", # error suffix in exception names
+    "RUF067", # no code in __init__ *shrug*
 ]
 allowed-confusables = ["‐", "–"]
 
