### 2026-01-19 #9294 @Nikoleta-v3
@alexfikl, could you please:

- Let me know whether I should go ahead and use `v2025.1` as the version, or if you would prefer to create and use a new release.
- Archive the reviewed software on Zenodo or a similar service (e.g., Figshare or an institutional repository).
- Ensure that the archival deposit (e.g., on Zenodo) includes the correct metadata, including:
  - The correct title (matching the paper title)
  - An accurate author list
  - Optional: ORCIDs for each author

Finally, please share the DOI of the archived version here.
