### 2026-01-19 #9294 @alexfikl
> @alexfikl, could you please:

Will get right on this, hopefully in the next couple of days, and keep you updated.

> Ensure that the archival deposit (e.g., on Zenodo) includes the correct metadata, including:

Since we have used a separate branch for the JOSS submission, does this mean we have to create a separate Zenodo deposit for it that matches the paper?
