### 2026-01-19 4620999
chore: add ORCiD to CITATION.cff
