+++ CITATION.cff
@@ -6,10 +6,13 @@ authors:
   orcid: "https://orcid.org/0000-0003-1228-519X"
 - family-names: "Fikl"
   given-names: "Alex"
+  orcid: "https://orcid.org/0000-0002-0552-5936"
 - family-names: "Wei"
   given-names: "Xiaoyu"
+  orcid: "https://orcid.org/0000-0001-7063-7865"
 - family-names: "Gibson"
   given-names: "Thomas"
+  orcid: "https://orcid.org/0000-0002-7978-6848"
 - family-names: "Alvey-Blanco"
   given-names: "Addison"
 - family-names: "Wala"
