### 2026-01-19 af805e1
chore: bump version in CITATION.cff
