+++ CITATION.cff
@@ -19,8 +19,8 @@ authors:
   given-names: "Matt"
 
 title: "modepy"
-version: 2024.1
-doi: 10.5281/zenodo.11105052
-date-released: 2024-05-02
+version: 2025.1
+doi: 10.5281/zenodo.14624221
+date-released: 2025-01-09
 url: "https://github.com/inducer/modepy"
 license: MIT
