### 2026-01-19 dc9b4c9
Add automated release Github Action
