+++ pyproject.toml
@@ -6,7 +6,7 @@ requires = [
 
 [project]
 name = "modepy"
-version = "2025.2"
+version = "2026.1"
 description = "Modes and nodes for high-order methods"
 readme = "README.rst"
 license = "MIT"
