### 2026-01-20 #9294 @Nikoleta-v3
The Zenodo archive should correspond to the released version of the software. It can be the master branch, but it needs to be the version that the paper is “capturing”.
