### 2026-01-20 #9294 @alexfikl
> The Zenodo archive should correspond to the released version of the software. It can be the master branch, but it needs to be the version that the paper is “capturing”.

Thanks for the clarification! Some progress:

1. We made a new v2026.1 release to match the paper: https://github.com/inducer/modepy/releases/tag/v2026.1. We'd prefer using this one, of course. :grin:
2. We archived v2026.1 on Zenodo:
	- URL: https://zenodo.org/records/18304836
	- DOI: https://doi.org/10.5281/zenodo.18304836
	- We've also added the ORCiDs to the authors and they should match the paper.

Let me know if there are any missing pieces.
