### 2026-01-20 #9294 @Nikoleta-v3
@alexfikl thank you for making the new release! Regarding the Zenodo archive, could you please make sure that the names are exactly the same, including umlauts (ö) and any middle names?
