### 2026-01-20 #9294 @alexfikl
> could you please make sure that the names are exactly the same, including umlauts (ö) and any middle names?

Will do! I think the current version just matches the names on GitHub.
