### 2026-01-20 26ea75f
chore: update names in CITATION.cff to match paper
