+++ CITATION.cff
@@ -5,7 +5,7 @@ authors:
   given-names: "Andreas"
   orcid: "https://orcid.org/0000-0003-1228-519X"
 - family-names: "Fikl"
-  given-names: "Alex"
+  given-names: "Alexandru"
   orcid: "https://orcid.org/0000-0002-0552-5936"
 - family-names: "Wei"
   given-names: "Xiaoyu"
