### 2026-01-20 #9294 @alexfikl
> could you please make sure that the names are exactly the same, including umlauts (ö) and any middle names?

Would it also be acceptable to update the names in the paper? I'm mostly worried that changing the names for `modepy` will make the authors' other projects less discoverable on Zenodo (because clicking on the name won't show them anymore).
