### 2026-01-20 #9294 @Nikoleta-v3
@alexfikl can you not just edit the Zenodo archive? I’m guessing it was created using the GitHub–Zenodo integration, but aren’t you able to edit the existing Zenodo release directly? There’s no need to change the names in the repository itself.

Also, regarding the names in the paper, I would suggest always using the names as they appear in the authors’ other publications, if available.
