### 2026-01-20 f55c19a
feat(paper): update names to match code + Zenodo archive
