### 2026-01-20 #9294 @alexfikl
> [@alexfikl](https://github.com/alexfikl) can you not just edit the Zenodo archive? I’m guessing it was created using the GitHub–Zenodo integration, but aren’t you able to edit the existing Zenodo release directly? There’s no need to change the names in the repository itself.

Yep, that's what we ended up doing, thanks!

> Also, regarding the names in the paper, I would suggest always using the names as they appear in the authors’ other publications, if available.

@Nikoleta-v3  With this in mind, we went ahead and modified some of the names in the paper + some of the names on Zenodo to better match previous publications. I was a bit overzealous initially with the middle names, apparently 🤦‍♂️.

Everyone should match now.
