### 2026-01-21 #9294 @Nikoleta-v3
> @Nikoleta-v3 With this in mind, we went ahead and modified some of the names in the paper + some of the names on Zenodo to better match previous publications. I was a bit overzealous initially with the middle names, apparently 🤦‍♂️.

😆

Thank you @alexfikl! What about  [Wala, Matt](https://zenodo.org/search?q=metadata.creators.person_or_org.name:%22Wala,+Matt%22)? I just realized they are on the Zenodo list of authors but not on the paper.
