### 2026-01-21 #9294 @alexfikl
> Thank you [@alexfikl](https://github.com/alexfikl)! What about [Wala, Matt](https://zenodo.org/search?q=metadata.creators.person_or_org.name:%22Wala,+Matt%22)? I just realized they are on the Zenodo list of authors but not on the paper.

We didn't add Matt to the paper because he just has a few non-code contributions (see [here](https://github.com/inducer/modepy/commits?author=mattwala)). Is the policy that all contributors be included?
