### 2026-01-21 #9294 @inducer
I can remove him from Zenodo too if it's important. Not a big deal to anyone involved, I think (hope?).

cc @mattwala
