### 2026-01-22 #9294 @Nikoleta-v3
Not an issue at all, please do go ahead and remove them from the Zenodo archieve. For authorship, either substantial code contributions or intellectual/design contributions should generally be considered sufficient. From what I can tell from the commit history (https://github.com/inducer/modepy/commits?author=mattwala), these appear to be relatively minor additions.

In any case, if @matt-graham agrees that he does not wish to be included on the author list, then please feel free to proceed and remove the name from Zenodo @inducer @alexfikl
