### 2026-01-22 4cb04e1
Update CITATION.cff for JOSS
