### 2026-01-22 #9294 @inducer
I've removed Matt from the [Zenodo record](https://zenodo.org/records/18304836), and https://github.com/inducer/modepy/pull/155 will update `CITATION.cff` in the source repo accordingly.
