### 2026-01-27 #9294 @Nikoleta-v3
@alexfikl thank you for the clarification. Given the commit history and your explanation of the situation, this is sufficient for this case. Obviously, if the contributions were more ambiguous, we would need a proper reply from them, but given the specific situation, I’m going to go ahead and accept the current list of authors.
