### 2026-01-27 #9294 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
