### 2026-01-27 #9294 @alexfikl
> I know this hasn’t been a linear process, with reviewers dropping out and other issues [@alexfikl](https://github.com/alexfikl) . So I just wanted to say thank you for your patience throughout the process 😄 🙏🏻

Not at all, it's been quite pleasant for our first JOSS submission! Thank you for all the help and guidance in getting this ready :grin:
