### 2026-01-27 #9294 @crvernon
@alexfikl - I just need you to make the following changes before I accept:

In your archive:

- [x] The title of your archive must match the title of your paper exactly.  Please edit the metadata in your [archive]([10.5281/zenodo.18304836](https://doi.org/10.5281/zenodo.18304836)) so that the title appears as "modepy: Basis Functions, Interpolation, and Quadrature (not just) for Finite Elements" which is the current title of your paper.  No new release is required for this modification.

In your paper:

- [x] LINE 75:  You state "...A Basis objects is..."  do you mean "...A Basis object is..." where "objects" should be "objects"?  If this is the desired phrasing, then ignore this comment.
- [x] LINE 77:  The in-text citation "...basis from (Dubiner, 1991)..." should be formatted as "...basis from Dubiner (1991)..."  This can be achieved by simply removing the brackets from your current in-text reference to the BibTeX target.
- [x] LINE 86:  The in-text citation "...described in (Isaac, 2020)..." should be formatted as "...described in Isaac (2020)..."
- [x] LINE 147-149:  The Meurer et al. reference has a DOI listed here:  https://peerj.com/articles/cs-103/ (see the "Cite this as" block under the title.  Here is the DOI to apply to this record:  `https://doi.org/10.7717/peerj-cs.103`

Let me know when these have been addressed and I'll accept after verification.  Thank you!
