### 2026-01-27 eaca9da
feat(paper): address additional comments from review
