+++ doc/paper/paper.bib
@@ -250,6 +250,7 @@
   year          = 2017,
   journal       = {PeerJ Computer Science},
   publisher     = {PeerJ Inc.},
+  doi           = {10.7717/peerj-cs.103},
   volume        = 3,
   pages         = {e103},
 }
