### 2026-01-27 #9294 @alexfikl
> In your archive:

Will update this later today.

> In your paper:

These should all be fixed. Thanks for the DOI on the `sympy` paper!
