### 2026-01-27 #9294 @inducer
Updated: https://zenodo.org/records/18304836
